Improve testing equivalence of metadata association graphs
Previously, two data file - metadata file association graphs were compared simply by ensuring that all files in each list were present in the other. This however failed to account for the fact that in many circumstances these lists must be interpreted in an ordered way, and so ensuring that their orders are equivalent is important. The lists however cannot simply be tested for equivalence. A ruleset may permit a single data file to load key-value metadata from multiple JSON files that are equivalent in both filesystem directory location and number of entities. In this scenario, the order within that set of equidistant metadats files is not of consequence in a test of equivalence. This commit replaces the earlier placeholder implementation with this more robust algorithm.